requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
joblib==1.4.2
nltk==3.9.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
transformers==4.51.2
torchaudio==2.5.0
torchvision==0.20.0
torch==2.6.0
huggingface_hub[hf_xet]==0.17.2 
